refactor(auth): one session mint, per-provider SSO, four readable units
Unify: finalizeLogin, Refresh, and SwitchWorkspace each carried their own
access+refresh cookie block — issueSessionCookies is now the single place
a web session is minted, with the two expiry shapes (fresh session vs.
rotation inheriting the absolute lifetime) passed explicitly. The 230-line
IDP switch splits into oauth2UserInfo, oidcUserInfo, and ldapUserInfo, so
getOrCreateUserWithIDP reads as: resolve provider, fetch identity,
resolve email, provision.

Reorganize: the package now reads as four units — auth_service.go holds
the session core (Login pipeline, Signup, Refresh, SwitchWorkspace,
tokens/cookies, workspace resolution) at 1233 lines, with SSO in
auth_service_idp.go, emailed codes in auth_service_email_code.go, and the
lockout in auth_service_lockout.go, each behind a one-line scope header.
Login's step comments are renumbered. Pure moves and extractions; no
behavior change.
